llm_batch_sg/src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
databricks-connect==15.4.2
databricks-sdk==0.34.0
databricks-sql-connector==3.4.0
datasets==3.0.1
dill==0.3.8
et-xmlfile==1.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
huggingface-hub==0.25.2
idna==3.10
lz4==4.3.3
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.0
protobuf==5.28.2
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
SQLAlchemy==2.0.36
thrift==0.20.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.15.3
